+++ CHANGES.md
@@ -21,6 +21,8 @@
     - @voluntas
 - [UPDATE] asmjit を `0dd16b0a98ae1da48563c9cc62f757a9e6bbe9b6` に上げる
     - @voluntas
+- [ADD] DataChannel シグナリングに対応
+    - @melpon
 
 ## 2021.1.2